% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_replicate}
\alias{add_replicate}
\title{Add replicate pattern in survey design with weight and replicate files}
\usage{
add_replicate(
  weight,
  replicate_pattern,
  replicate_path = NULL,
  replicate_id = NULL,
  replicate_type
)
}
\arguments{
\item{weight}{String with name of weight in survey}

\item{replicate_pattern}{String with replicate pattern for columns names regex expression}

\item{replicate_path}{String with replicate file path}

\item{replicate_id}{String with replicate ID use to join with survey}

\item{replicate_type}{String with replicate type, e.g. "bootstrap", "jackknife"}
}
\value{
List of replicate parameters. It will be used to create a replicate object in the survey design.
}
\description{
Add replicate pattern in survey design with weight and replicate files
}
\examples{
add_replicate(
  weight = "W",
  replicate_pattern = "wr[0-9]+",
  replicate_path = here::here(
    "example-data",
    "ech",
    "ech_2023",
    "pesos replicados Bootstrap anual 2023.xlsx"
  ),
  replicate_id = c("ID" = "ID"),
  replicate_type = "bootstrap"
)

}
\keyword{utils}
