% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_weight}
\alias{add_weight}
\title{Add Weight time pattern}
\usage{
add_weight(monthly = NULL, annual = NULL, quarterly = NULL, biannual = NULL)
}
\arguments{
\item{monthly}{Weight monthly}

\item{annual}{Weight annual}

\item{quarterly}{Weight quarterly}

\item{biannual}{Weight biannual}
}
\description{
Add Weight time pattern
}
\keyword{utils}
