% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_comment_recipe}
\alias{api_comment_recipe}
\title{Add a comment to a recipe}
\usage{
api_comment_recipe(id, text)
}
\arguments{
\item{id}{Recipe ID}

\item{text}{Comment text (max 2000 characters)}
}
\value{
Invisibly, the API response.
}
\description{
Post a text comment on a recipe. Requires authentication.
}
\examples{
\dontrun{
api_comment_recipe("r_1739654400_742", "Great recipe!")
}
}
\seealso{
Other api-comments: 
\code{\link{api_comment_workflow}()},
\code{\link{api_delete_comment}()},
\code{\link{api_get_recipe_comments}()},
\code{\link{api_get_workflow_comments}()}
}
\concept{api-comments}
