% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_comment_workflow}
\alias{api_comment_workflow}
\title{Add a comment to a workflow}
\usage{
api_comment_workflow(id, text)
}
\arguments{
\item{id}{Workflow ID}

\item{text}{Comment text (max 2000 characters)}
}
\value{
Invisibly, the API response.
}
\description{
Post a text comment on a workflow. Requires authentication.
}
\examples{
\dontrun{
api_comment_workflow("w_1739654400_123", "Very useful workflow!")
}
}
\seealso{
Other api-comments: 
\code{\link{api_comment_recipe}()},
\code{\link{api_delete_comment}()},
\code{\link{api_get_recipe_comments}()},
\code{\link{api_get_workflow_comments}()}
}
\concept{api-comments}
