% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_download_recipe}
\alias{api_download_recipe}
\title{Increment recipe download counter}
\usage{
api_download_recipe(id)
}
\arguments{
\item{id}{Recipe ID}
}
\description{
Increment recipe download counter
}
\keyword{internal}
