% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_download_workflow}
\alias{api_download_workflow}
\title{Increment workflow download counter}
\usage{
api_download_workflow(id)
}
\arguments{
\item{id}{Workflow ID}
}
\description{
Increment workflow download counter
}
\keyword{internal}
