% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_get_anda_variables}
\alias{api_get_anda_variables}
\title{Get ANDA variable metadata from the API}
\usage{
api_get_anda_variables(survey_type = "ech", var_names = NULL)
}
\arguments{
\item{survey_type}{Character survey type (default "ech")}

\item{var_names}{Character vector of variable names. If NULL, returns all.}
}
\value{
A list of variable metadata objects
}
\description{
Get ANDA variable metadata from the API
}
\examples{
\donttest{
api_get_anda_variables("ech", c("pobpcoac", "e27"))
}
}
\seealso{
Other anda: 
\code{\link{anda_download_microdata}()},
\code{\link{anda_variables}()}
}
\concept{anda}
