% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_get_recipe}
\alias{api_get_recipe}
\title{Get a single recipe by ID}
\usage{
api_get_recipe(id)
}
\arguments{
\item{id}{Recipe ID}
}
\value{
Recipe object or NULL
}
\description{
Get a single recipe by ID
}
