% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_get_recipe_comments}
\alias{api_get_recipe_comments}
\title{Get comments for a recipe}
\usage{
api_get_recipe_comments(id)
}
\arguments{
\item{id}{Recipe ID}
}
\value{
List of comment objects.
}
\description{
List all comments on a recipe, sorted by creation date.
}
\examples{
\dontrun{
api_get_recipe_comments("r_1739654400_742")
}
}
\seealso{
Other api-comments: 
\code{\link{api_comment_recipe}()},
\code{\link{api_comment_workflow}()},
\code{\link{api_delete_comment}()},
\code{\link{api_get_workflow_comments}()}
}
\concept{api-comments}
