% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_get_recipe_dependents}
\alias{api_get_recipe_dependents}
\title{Get recipes that depend on a recipe}
\usage{
api_get_recipe_dependents(id)
}
\arguments{
\item{id}{Recipe ID}
}
\value{
List of recipe summaries (id, name, user).
}
\description{
Find all recipes whose \code{depends_on_recipes} field
references the given recipe ID (backlinks).
}
\examples{
\dontrun{
api_get_recipe_dependents("r_1739654400_742")
}
}
\seealso{
Other api-recipes: 
\code{\link{api_get_recipe}()},
\code{\link{api_list_recipes}()},
\code{\link{api_publish_recipe}()}
}
\concept{api-recipes}
