% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_get_workflow}
\alias{api_get_workflow}
\title{Get a single workflow by ID}
\usage{
api_get_workflow(id)
}
\arguments{
\item{id}{Workflow ID}
}
\value{
RecipeWorkflow object or NULL
}
\description{
Get a single workflow by ID
}
