% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_list_workflows}
\alias{api_list_workflows}
\title{List workflows from API}
\usage{
api_list_workflows(
  user = NULL,
  edition = NULL,
  survey_type = NULL,
  recipe_id = NULL,
  limit = 100
)
}
\arguments{
\item{user}{Filter by user email}

\item{edition}{Filter by edition}

\item{survey_type}{Filter by survey type}

\item{recipe_id}{Filter by recipe ID}

\item{limit}{Maximum results}
}
\value{
List of RecipeWorkflow objects
}
\description{
List workflows from API
}
