% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_logout}
\alias{api_logout}
\title{Logout}
\usage{
api_logout()
}
\value{
Invisibly, NULL.
}
\description{
Clear the stored API token from memory and the environment.
}
\examples{
\dontrun{
api_logout()
}
}
