% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_publish_recipe}
\alias{api_publish_recipe}
\title{Publish recipe to API}
\usage{
api_publish_recipe(recipe)
}
\arguments{
\item{recipe}{Recipe object}
}
\value{
API response
}
\description{
Publish recipe to API
}
