% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_publish_workflow}
\alias{api_publish_workflow}
\title{Publish workflow to API}
\usage{
api_publish_workflow(workflow)
}
\arguments{
\item{workflow}{RecipeWorkflow object}
}
\value{
API response
}
\description{
Publish workflow to API
}
