% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_track_download}
\alias{api_track_download}
\title{Increment download counter (generic)}
\usage{
api_track_download(type, id)
}
\arguments{
\item{type}{\code{"recipe"} or \code{"workflow"}}

\item{id}{Object ID}
}
\description{
Increment download counter (generic)
}
\keyword{internal}
