% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step.R
\name{bake_ast_step}
\alias{bake_ast_step}
\title{Bake AST-based step}
\usage{
bake_ast_step(svy, step)
}
\arguments{
\item{svy}{Survey object}

\item{step}{AST step to bake}
}
\value{
Modified survey object
}
\description{
Bake AST-based step
}
\keyword{internal}
