% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{bake_recipe}
\alias{bake_recipe}
\title{Bake a recipe}
\usage{
bake_recipe(recipe, svy)
}
\arguments{
\item{recipe}{A Recipe object}

\item{svy}{A Survey object}
}
\value{
A Survey object
}
\description{
Bake a recipe
}
\keyword{Recipes}
\keyword{Survey}
\keyword{methods}
