% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{bake_recipes}
\alias{bake_recipes}
\title{Bake recipes}
\usage{
bake_recipes(svy, recipes)
}
\arguments{
\item{svy}{Survey object}

\item{recipes}{List of recipes}
}
\value{
Survey object
}
\description{
Bake recipes
}
\keyword{Surveymethods}
