% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{cat_design_type}
\alias{cat_design_type}
\title{cat_design_type}
\usage{
cat_design_type(self, design_name)
}
\arguments{
\item{self}{Object of class Survey}

\item{design_name}{Name of design}
}
\value{
Character string describing the design type, or "None".
}
\description{
Cast design type from survey
}
\examples{
\dontrun{
svy <- load_survey("data.csv", svy_type = "ech", svy_edition = "2023")
cat_design_type(svy, "annual")
}
}
\keyword{Surveymethods}
