% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{cat_recipes}
\alias{cat_recipes}
\title{cat_recipes}
\usage{
cat_recipes(self)
}
\arguments{
\item{self}{Object of class Survey}
}
\value{
Character string listing recipe names, or "None".
}
\description{
Cast recipes from survey
}
\examples{
\dontrun{
svy <- load_survey("data.csv", svy_type = "ech", svy_edition = "2023")
cat_recipes(svy)
}
}
\keyword{internal}
\keyword{survey}
