% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_engine.R
\name{default_engine}
\alias{default_engine}
\title{default_engine}
\usage{
default_engine(.engine = "data.table")
}
\arguments{
\item{.engine}{Character vector with the name of the engine}
}
\description{
Set default engine
}
\keyword{engine}
