% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anda.R
\name{.anda_select_resource}
\alias{.anda_select_resource}
\title{Select resources matching the requested type}
\usage{
.anda_select_resource(resources, resource, edition)
}
\arguments{
\item{resources}{data.frame from .anda_parse_resources}

\item{resource}{Character resource type}

\item{edition}{Character edition year}
}
\value{
data.frame subset of matching resources
}
\description{
Select resources matching the requested type
}
\keyword{internal}
