% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{encoding_recipe}
\alias{encoding_recipe}
\title{Encoding and decoding recipes}
\usage{
encoding_recipe(recipe)
}
\arguments{
\item{recipe}{A Recipe object}
}
\value{
A Recipe object
}
\description{
Encoding and decoding recipes
}
