% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{evaluate_ast_node}
\alias{evaluate_ast_node}
\title{Evaluate single AST node}
\usage{
evaluate_ast_node(node, data, env)
}
\arguments{
\item{node}{AST node to evaluate}

\item{data}{Data for evaluation}

\item{env}{Environment for functions}
}
\value{
Evaluation result
}
\description{
Evaluate single AST node
}
\keyword{internal}
