% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_cv}
\alias{evaluate_cv}
\title{Evaluate estimation with Coeficient of Variation}
\usage{
evaluate_cv(cv)
}
\arguments{
\item{cv}{Coeficient of Variation}
}
\value{
character
}
\description{
Evaluate estimation with Coeficient of Variation
}
\keyword{utils}
