% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_parser.R
\name{expand_loop_body}
\alias{expand_loop_body}
\title{Substitute loop variable in body lines}
\usage{
expand_loop_body(loopvar, values, body)
}
\arguments{
\item{loopvar}{Loop variable name}

\item{values}{Character vector of values}

\item{body}{Character vector of body lines}
}
\value{
Character vector of expanded lines
}
\description{
Substitute loop variable in body lines
}
\keyword{internal}
