% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_parser.R
\name{expand_numlist}
\alias{expand_numlist}
\title{Expand a numlist specification into numeric values}
\usage{
expand_numlist(spec)
}
\arguments{
\item{spec}{Numlist string like "1/14" or "81/89"}
}
\value{
Character vector of values
}
\description{
Expand a numlist specification into numeric values
}
\keyword{internal}
