% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_mappings.R
\name{expand_var_range}
\alias{expand_var_range}
\title{Expand a STATA variable range to individual variable names}
\usage{
expand_var_range(spec)
}
\arguments{
\item{spec}{Variable specification (may contain ranges with -)}
}
\value{
Character vector of individual variable names
}
\description{
STATA allows variable ranges like suma1-suma4 which means
suma1 suma2 suma3 suma4. This function detects and expands
such ranges by incrementing the trailing numeric suffix.
}
\keyword{internal}
