% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelSurvey.R
\name{extract_surveys}
\alias{extract_surveys}
\title{Extract surveys}
\usage{
extract_surveys(
  RotativePanelSurvey,
  index = NULL,
  monthly = NULL,
  annual = NULL,
  quarterly = NULL,
  biannual = NULL,
  use.parallel = FALSE
)
}
\arguments{
\item{RotativePanelSurvey}{A RotativePanelSurvey object}

\item{index}{An integer}

\item{monthly}{A vector of integers}

\item{annual}{A vector of integers}

\item{quarterly}{A vector of integers}

\item{biannual}{A vector of integers}

\item{use.parallel}{A logical}
}
\value{
A list of surveys
}
\description{
Extract surveys
}
\keyword{RotativePanelSurvey}
\keyword{Surveymethods}
