% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeAPI.R
\name{get_backend}
\alias{get_backend}
\title{Get recipe backend}
\usage{
get_backend()
}
\value{
RecipeBackend object
}
\description{
Returns the currently configured recipe backend.
Defaults to "api" if not configured.
}
\examples{
\dontrun{
backend <- get_backend()
backend$search("labor")
}
}
