% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(svy)
}
\arguments{
\item{svy}{A Survey object.}
}
\value{
The survey data.
}
\description{
Retrieve the data from a Survey object.
}
\keyword{Surveymethods}
