% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_edition}
\alias{get_edition}
\title{get_edition}
\usage{
get_edition(svy)
}
\arguments{
\item{svy}{A Survey object.}
}
\value{
The survey edition.
}
\description{
Retrieve the edition of a Survey object.
}
\keyword{Surveymethods}
