% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_info_weight}
\alias{get_info_weight}
\title{get_info_weight}
\usage{
get_info_weight(svy)
}
\arguments{
\item{svy}{A Survey object.}
}
\value{
A string with detailed weight information.
}
\description{
Retrieve detailed information about the weights of a Survey object.
}
\keyword{Surveymethods}
