% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_metadata}
\alias{get_metadata}
\title{get_metadata}
\usage{
get_metadata(self)
}
\arguments{
\item{self}{Object of class Survey}
}
\value{
NULL (called for side effect: prints metadata to console).
}
\description{
Get metadata from survey
}
\examples{
dt <- data.table::data.table(id = 1:5, age = c(25, 30, 45, 50, 60), w = rep(1, 5))
svy <- Survey$new(
  data = dt, edition = "2023", type = "ech",
  psu = NULL, engine = "data.table", weight = add_weight(annual = "w")
)
get_metadata(svy)
}
\keyword{Surveymethods}
