% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_steps}
\alias{get_steps}
\title{get_steps}
\usage{
get_steps(svy)
}
\arguments{
\item{svy}{Survey object}
}
\value{
List of Step objects
}
\description{
Get steps from survey
}
\examples{
dt <- data.table::data.table(id = 1:5, age = c(25, 30, 45, 50, 60), w = rep(1, 5))
svy <- Survey$new(
  data = dt, edition = "2023", type = "ech",
  psu = NULL, engine = "data.table", weight = add_weight(annual = "w")
)
svy <- step_compute(svy, age2 = age * 2)
get_steps(svy) # list of Step objects
}
\keyword{Steps}
\keyword{Survey}
\keyword{methods}
