% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_type}
\alias{get_type}
\title{get_type}
\usage{
get_type(svy)
}
\arguments{
\item{svy}{A Survey object.}
}
\value{
The survey type.
}
\description{
Retrieve the type of a Survey object.
}
\keyword{Surveymethods}
