% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_weight}
\alias{get_weight}
\title{get_weight}
\usage{
get_weight(svy, estimation_type = 1:length(svy$weight))
}
\arguments{
\item{svy}{A Survey object.}

\item{estimation_type}{The type of estimation (default: all weights).}
}
\value{
The weight(s) of the survey.
}
\description{
Retrieve the weight(s) of a Survey object.
}
\keyword{Surveymethods}
