% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowBackend.R
\name{get_workflow_backend}
\alias{get_workflow_backend}
\title{Get workflow backend}
\usage{
get_workflow_backend()
}
\value{
WorkflowBackend object
}
\description{
Returns the currently configured workflow backend.
Defaults to "local" if not configured.
}
\examples{
\dontrun{
backend <- get_workflow_backend()
backend$search("labor")
}
}
