% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{group_dates}
\alias{group_dates}
\title{Group dates}
\usage{
group_dates(dates, type = c("monthly", "quarterly", "biannual"))
}
\arguments{
\item{dates}{Dates}

\item{type}{Type}
}
\value{
Group
}
\description{
Group dates
}
\keyword{utils}
