% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{has_steps}
\alias{has_steps}
\title{Check if survey has steps}
\usage{
has_steps(svy)
}
\arguments{
\item{svy}{A Survey or RotativePanelSurvey object.}
}
\value{
Logical.
}
\description{
Check if survey has steps
}
\examples{
svy <- survey_empty(type = "test", edition = "2023")
has_steps(svy) # FALSE
}
\seealso{
Other survey-objects: 
\code{\link{Survey}},
\code{\link{cat_design}()},
\code{\link{cat_design_type}()},
\code{\link{get_data}()},
\code{\link{get_metadata}()},
\code{\link{has_design}()},
\code{\link{has_recipes}()},
\code{\link{is_baked}()},
\code{\link{set_data}()},
\code{\link{survey_empty}()},
\code{\link{survey_to_data.table}()},
\code{\link{survey_to_data_frame}()},
\code{\link{survey_to_tibble}()}
}
\concept{survey-objects}
