% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_parser.R
\name{join_continuation_lines}
\alias{join_continuation_lines}
\title{Join STATA continuation lines}
\usage{
join_continuation_lines(lines)
}
\arguments{
\item{lines}{Character vector of source lines}
}
\value{
Character vector with continuations joined
}
\description{
STATA uses /// for line continuation. This function joins
continued lines into single commands.
}
\keyword{internal}
