% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_tidy_api.R
\name{list_recipes}
\alias{list_recipes}
\title{List all recipes}
\usage{
list_recipes()
}
\value{
List of all Recipe objects.
}
\description{
List all recipes from the active backend.
}
\examples{
set_backend("local", path = tempfile(fileext = ".json"))
all <- list_recipes()
length(all)

}
\seealso{
\code{\link{search_recipes}}, \code{\link{filter_recipes}}
}
