% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_survey.R
\name{load_panel_survey}
\alias{load_panel_survey}
\title{Load panel survey}
\usage{
load_panel_survey(
  path_implantation,
  path_follow_up,
  svy_type,
  svy_weight_implantation,
  svy_weight_follow_up
)
}
\arguments{
\item{path_implantation}{Path to the implantation survey}

\item{path_follow_up}{Path to the follow-up survey}

\item{svy_type}{Type of survey}

\item{svy_weight_implantation}{Weight of the implantation survey}

\item{svy_weight_follow_up}{Weight of the follow-up survey}
}
\value{
RotativePanelSurvey object
}
\description{
Load panel survey
}
