% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_survey_example}
\alias{load_survey_example}
\title{Load survey example}
\usage{
load_survey_example(svy_type, svy_edition)
}
\arguments{
\item{svy_type}{Survey type}

\item{svy_edition}{Survey edition}
}
\description{
Load survey example
}
\keyword{utils}
