% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{metadata_recipe}
\alias{metadata_recipe}
\title{Metadata for Recipe}
\usage{
metadata_recipe()
}
\value{
A character vector of metadata field names.
}
\description{
Returns the metadata fields required for a recipe.
}
\keyword{internal}
