% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{optimize_node}
\alias{optimize_node}
\title{Optimize individual AST node}
\usage{
optimize_node(node, data = NULL)
}
\arguments{
\item{node}{AST node to optimize}

\item{data}{Optional data context}
}
\value{
Optimized node
}
\description{
Optimize individual AST node
}
\keyword{internal}
