% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{parse_call_ast}
\alias{parse_call_ast}
\title{Parse function call into AST}
\usage{
parse_call_ast(call_expr, env)
}
\arguments{
\item{call_expr}{A call expression}

\item{env}{Environment for resolution}
}
\value{
AST node for function call
}
\description{
Parse function call into AST
}
\keyword{internal}
