% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_mappings.R
\name{parse_destring_args}
\alias{parse_destring_args}
\title{Parse a STATA destring command}
\usage{
parse_destring_args(args, options = NULL)
}
\arguments{
\item{args}{Arguments string after "destring"}

\item{options}{Options string}
}
\value{
List with var_name, replace (logical), force (logical), gen_var
}
\description{
Parse a STATA destring command
}
\keyword{internal}
