% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_mappings.R
\name{parse_egen_args}
\alias{parse_egen_args}
\title{Parse a STATA egen command}
\usage{
parse_egen_args(args, by_group = NULL, options = NULL)
}
\arguments{
\item{args}{Arguments string after "egen"}

\item{by_group}{By-group from bysort prefix or by() option}

\item{options}{Options string}
}
\value{
List with var_name, func, func_arg, by_group
}
\description{
Parse a STATA egen command
}
\keyword{internal}
