% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_mappings.R
\name{parse_gen_args}
\alias{parse_gen_args}
\title{Parse a STATA gen command into variable name and expression}
\usage{
parse_gen_args(args)
}
\arguments{
\item{args}{The arguments part of a gen command (after "gen")}
}
\value{
List with var_name and expr
}
\description{
Parse a STATA gen command into variable name and expression
}
\keyword{internal}
