% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{parse_literal_ast}
\alias{parse_literal_ast}
\title{Parse literal into AST}
\usage{
parse_literal_ast(literal_expr)
}
\arguments{
\item{literal_expr}{A literal value}
}
\value{
AST node for literal
}
\description{
Parse literal into AST
}
\keyword{internal}
