% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_mappings.R
\name{parse_mvencode_args}
\alias{parse_mvencode_args}
\title{Parse a STATA mvencode command}
\usage{
parse_mvencode_args(args, options = NULL)
}
\arguments{
\item{args}{Arguments string after "mvencode"}

\item{options}{Options string}
}
\value{
List with var_names and mv_value
}
\description{
Parse a STATA mvencode command
}
\keyword{internal}
