% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stata_mappings.R
\name{parse_replace_args}
\alias{parse_replace_args}
\title{Parse a STATA replace command into variable name and expression}
\usage{
parse_replace_args(args)
}
\arguments{
\item{args}{The arguments part of a replace command}
}
\value{
List with var_name and expr
}
\description{
Parse a STATA replace command into variable name and expression
}
\keyword{internal}
