% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{parse_symbol_ast}
\alias{parse_symbol_ast}
\title{Parse symbol into AST}
\usage{
parse_symbol_ast(symbol_expr, env)
}
\arguments{
\item{symbol_expr}{A symbol expression}

\item{env}{Environment for resolution}
}
\value{
AST node for symbol
}
\description{
Parse symbol into AST
}
\keyword{internal}
