% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{publish_recipe}
\alias{publish_recipe}
\title{Publish Recipe}
\usage{
publish_recipe(recipe)
}
\arguments{
\item{recipe}{A Recipe object.}
}
\value{
A JSON object containing the API response.
}
\description{
Publishes a Recipe object to the API.
}
\details{
This function sends a Recipe object to the API for publication.
}
\examples{
\dontrun{
# Example of publishing a Recipe object to the API
publish_recipe(recipe_obj)
}
}
\keyword{utils}
