% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_tidy_api.R
\name{rank_workflows}
\alias{rank_workflows}
\title{Rank workflows by downloads}
\usage{
rank_workflows(n = NULL)
}
\arguments{
\item{n}{Integer. Maximum number to return, or NULL for all.}
}
\value{
List of RecipeWorkflow objects sorted by downloads.
}
\description{
Get the top workflows ranked by download count.
}
\examples{
\dontrun{
top5 <- rank_workflows(n = 5)
}

}
\seealso{
\code{\link{search_workflows}}, \code{\link{filter_workflows}}
}
