% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{Recipe}
\alias{Recipe}
\title{Recipe Class}
\description{
This class represents a recipe for processing survey data. A recipe includes metadata such as name, edition, survey type, user, description, and a list of steps to be applied to the survey data.
}
\keyword{Recipes}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A string representing the name of the recipe.}

\item{\code{edition}}{A string indicating the edition of the survey associated with the recipe.}

\item{\code{survey_type}}{A string specifying the type of survey (e.g., "eaii", "ech").}

\item{\code{default_engine}}{A string indicating the default engine used for processing.}

\item{\code{depends_on}}{A list of dependencies required by the recipe.}

\item{\code{user}}{A string representing the user or creator of the recipe.}

\item{\code{description}}{A string providing a description of the recipe.}

\item{\code{id}}{A unique identifier for the recipe.}

\item{\code{steps}}{A list of steps to be applied to the survey data.}

\item{\code{doi}}{A string representing the DOI (Digital Object Identifier) of the recipe.}

\item{\code{bake}}{A logical value indicating whether the recipe should be baked (processed) immediately.}

\item{\code{topic}}{A string specifying the topic of the recipe.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Recipe-new}{\code{Recipe$new()}}
\item \href{#method-Recipe-clone}{\code{Recipe$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recipe-new"></a>}}
\if{latex}{\out{\hypertarget{method-Recipe-new}{}}}
\subsection{Method \code{new()}}{
Initializes a new instance of the Recipe class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recipe$new(
  name,
  edition,
  survey_type,
  default_engine,
  depends_on,
  user,
  description,
  steps,
  id,
  doi,
  topic
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{A string representing the name of the recipe.}

\item{\code{edition}}{A string indicating the edition of the survey associated with the recipe.}

\item{\code{survey_type}}{A string specifying the type of survey (e.g., "eaii", "ech").}

\item{\code{default_engine}}{A string indicating the default engine used for processing.}

\item{\code{depends_on}}{A list of dependencies required by the recipe.}

\item{\code{user}}{A string representing the user or creator of the recipe.}

\item{\code{description}}{A string providing a description of the recipe.}

\item{\code{steps}}{A list of steps to be applied to the survey data.}

\item{\code{id}}{A unique identifier for the recipe.}

\item{\code{doi}}{A string representing the DOI (Digital Object Identifier) of the recipe.}

\item{\code{topic}}{A string specifying the topic of the recipe.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Recipe-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Recipe-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Recipe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
