% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{recipe_to_json}
\alias{recipe_to_json}
\title{recipe to json}
\usage{
recipe_to_json(recipe)
}
\arguments{
\item{recipe}{A Recipe object}
}
\value{
A JSON object
}
\description{
recipe to json
}
\keyword{internal}
