% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{recode_with_ast}
\alias{recode_with_ast}
\title{Core AST recoding engine}
\usage{
recode_with_ast(
  svy,
  new_var,
  conditions,
  default_value = NA_character_,
  ast_params = list()
)
}
\arguments{
\item{svy}{Survey object}

\item{new_var}{Name of new variable to create}

\item{conditions}{List of condition formulas (LHS ~ RHS)}

\item{default_value}{Default value when no conditions match}

\item{ast_params}{AST configuration parameters}
}
\value{
Modified survey object with new recoded variable
}
\description{
This is the fundamental recoding engine that uses Abstract Syntax Tree
evaluation for all conditional logic in recoding operations.
}
\keyword{internal}
