% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{recode_with_ast_survey}
\alias{recode_with_ast_survey}
\title{AST survey recoding wrapper}
\usage{
recode_with_ast_survey(
  svy,
  new_var,
  conditions,
  default_value = NA_character_,
  use_copy = TRUE,
  to_factor = FALSE,
  ast_params = list()
)
}
\arguments{
\item{svy}{Survey object}

\item{new_var}{Name of new variable}

\item{conditions}{List of recoding conditions}

\item{default_value}{Default value for unmatched cases}

\item{use_copy}{Whether to copy survey object}

\item{to_factor}{Whether to convert result to factor}

\item{ast_params}{AST configuration parameters}
}
\value{
Modified survey object
}
\description{
Wrapper function that integrates AST recoding with survey object management
}
\keyword{internal}
