% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_tidy_api.R
\name{remove_category}
\alias{remove_category}
\title{Remove a category from a recipe}
\usage{
remove_category(recipe, name)
}
\arguments{
\item{recipe}{A Recipe object.}

\item{name}{Character. Category name to remove.}
}
\value{
The modified Recipe object.
}
\description{
Pipe-friendly function to remove a category from a Recipe by name.
}
\examples{
\dontrun{
my_recipe |>
  add_category("labor_market") |>
  remove_category("labor_market")
}

}
\seealso{
\code{\link{add_category}}
}
