% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_api_key}
\alias{set_api_key}
\title{Set API Key}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{API Key}
}
\description{
Set API Key
}
\examples{
\dontrun{
set_api_key("my-secret-key")
}
}
\keyword{utils}
