% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{set_edition}
\alias{set_edition}
\title{set_edition}
\usage{
set_edition(svy, new_edition, .copy = use_copy_default())
}
\arguments{
\item{svy}{A Survey object.}

\item{new_edition}{The new edition.}

\item{.copy}{Whether to return a copy of the Survey object (default: TRUE).}
}
\value{
The updated Survey object.
}
\description{
Set a new edition for a Survey object.
}
\keyword{Surveymethods}
