% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{set_type}
\alias{set_type}
\title{set_type}
\usage{
set_type(svy, new_type, .copy = use_copy_default())
}
\arguments{
\item{svy}{A Survey object.}

\item{new_type}{The new type.}

\item{.copy}{Whether to return a copy of the Survey object (default: TRUE).}
}
\value{
The updated Survey object.
}
\description{
Set a new type for a Survey object.
}
\keyword{Surveymethods}
