% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_use_copy}
\alias{set_use_copy}
\title{Set use_copy option}
\usage{
set_use_copy(use_copy)
}
\arguments{
\item{use_copy}{Use copy}
}
\description{
Set use_copy option
}
\examples{
set_use_copy(FALSE)
use_copy_default()
set_use_copy(TRUE)
use_copy_default()
}
