% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_tidy_api.R
\name{set_user_info}
\alias{set_user_info}
\title{Set user info on a recipe}
\usage{
set_user_info(recipe, user)
}
\arguments{
\item{recipe}{A Recipe object.}

\item{user}{A RecipeUser object.}
}
\value{
The modified Recipe object.
}
\description{
Pipe-friendly function to assign a \code{\link{RecipeUser}} to a Recipe.
}
\examples{
r <- recipe(
  name = "Example", user = "Test",
  svy = survey_empty(type = "ech", edition = "2023"),
  description = "Example recipe"
)
user <- recipe_user("Juan Perez", email = "juan@example.com")
r <- r |> set_user_info(user)

}
\seealso{
\code{\link{recipe_user}}
}
