% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{set_weight}
\alias{set_weight}
\title{set_weight}
\usage{
set_weight(svy, new_weight, .copy = use_copy_default())
}
\arguments{
\item{svy}{A Survey object.}

\item{new_weight}{The new weight(s).}

\item{.copy}{Whether to return a copy of the Survey object (default: TRUE).}
}
\value{
The updated Survey object.
}
\description{
Set new weight(s) for a Survey object.
}
\keyword{Surveymethods}
