% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_engine.R
\name{show_engines}
\alias{show_engines}
\title{show_engines}
\usage{
show_engines()
}
\value{
Character vector with the names of the available engines.
}
\description{
This function returns a list of available engines that can be used for loading surveys. The available engines are "data.table", "tidyverse", and "dplyr".
}
\keyword{engine}
