% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_compute}
\alias{step_compute}
\title{Step compute}
\usage{
step_compute(svy, ...)
}
\arguments{
\item{svy}{Survey object}

\item{...}{Expressions to compute}
}
\value{
Survey object
}
\description{
Step compute
}
