% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_rename}
\alias{step_rename}
\title{Rename variables in survey data (step)}
\usage{
step_rename(
  svy = survey_empty(),
  ...,
  mapping = NULL,
  use_copy = use_copy_default(),
  comment = "Rename variables",
  lazy = lazy_default(),
  record = TRUE
)
}
\arguments{
\item{svy}{A Survey or RotativePanelSurvey object}

\item{...}{Pairs in the form new_name = old_name (unquoted or character)}

\item{mapping}{A named character vector of the form \code{c(new_name = "old_name")}.}

\item{use_copy}{Whether to operate on a copy (default: use_copy_default())}

\item{comment}{Optional description for the step}

\item{lazy}{Logical, whether to delay execution.}

\item{record}{Logical, whether to record the step.}
}
\description{
Creates a step that renames variables in the survey data when baked.
}
