% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{Survey}
\alias{Survey}
\title{Survey Class}
\description{
This class represents a survey object with various attributes and methods to manage survey data, design, steps, recipes, and workflows.
}
\keyword{Survey}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The survey data.}

\item{\code{edition}}{The edition of the survey.}

\item{\code{type}}{The type´ of the survey.}

\item{\code{periodicity}}{The periodicity of the survey.}

\item{\code{default_engine}}{The default engine used for the survey.}

\item{\code{weight}}{The weight(s) associated with the survey.}

\item{\code{steps}}{A list of steps applied to the survey.}

\item{\code{recipes}}{A list of recipes associated with the survey.}

\item{\code{workflows}}{A list of workflows associated with the survey.}

\item{\code{design}}{The survey design object(s).}

\item{\code{active}}{A list of active bindings for dynamic properties of the survey.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Survey-new}{\code{Survey$new()}}
\item \href{#method-Survey-get_data}{\code{Survey$get_data()}}
\item \href{#method-Survey-get_edition}{\code{Survey$get_edition()}}
\item \href{#method-Survey-get_type}{\code{Survey$get_type()}}
\item \href{#method-Survey-set_data}{\code{Survey$set_data()}}
\item \href{#method-Survey-set_edition}{\code{Survey$set_edition()}}
\item \href{#method-Survey-set_type}{\code{Survey$set_type()}}
\item \href{#method-Survey-set_weight}{\code{Survey$set_weight()}}
\item \href{#method-Survey-print}{\code{Survey$print()}}
\item \href{#method-Survey-add_step}{\code{Survey$add_step()}}
\item \href{#method-Survey-add_recipe}{\code{Survey$add_recipe()}}
\item \href{#method-Survey-add_workflow}{\code{Survey$add_workflow()}}
\item \href{#method-Survey-bake}{\code{Survey$bake()}}
\item \href{#method-Survey-head}{\code{Survey$head()}}
\item \href{#method-Survey-str}{\code{Survey$str()}}
\item \href{#method-Survey-set_design}{\code{Survey$set_design()}}
\item \href{#method-Survey-update_design}{\code{Survey$update_design()}}
\item \href{#method-Survey-clone}{\code{Survey$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-new"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new Survey object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$new(
  data,
  edition,
  type,
  psu,
  engine,
  weight,
  design = NULL,
  steps = NULL,
  recipes = list()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The survey data.}

\item{\code{edition}}{The edition of the survey.}

\item{\code{type}}{The type of the survey.}

\item{\code{psu}}{Primary sampling unit (PSU).}

\item{\code{engine}}{The engine used for the survey.}

\item{\code{weight}}{The weight(s) for the survey.}

\item{\code{design}}{Optional survey design.}

\item{\code{steps}}{Optional list of steps.}

\item{\code{recipes}}{Optional list of recipes.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get the survey data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The survey data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-get_edition"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-get_edition}{}}}
\subsection{Method \code{get_edition()}}{
Get the edition of the survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$get_edition()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The survey edition.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-get_type"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-get_type}{}}}
\subsection{Method \code{get_type()}}{
Get the type of the survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$get_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The survey type.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-set_data"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-set_data}{}}}
\subsection{Method \code{set_data()}}{
Set the survey data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$set_data(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The new survey data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-set_edition"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-set_edition}{}}}
\subsection{Method \code{set_edition()}}{
Set the edition of the survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$set_edition(edition)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{edition}}{The new survey edition.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-set_type"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-set_type}{}}}
\subsection{Method \code{set_type()}}{
Set the type of the survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$set_type(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The new survey type.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-set_weight"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-set_weight}{}}}
\subsection{Method \code{set_weight()}}{
Set the weight(s) for the survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$set_weight(weight)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{weight}}{The new weight(s).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-print"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-print}{}}}
\subsection{Method \code{print()}}{
Print metadata of the survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-add_step"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-add_step}{}}}
\subsection{Method \code{add_step()}}{
Add a step to the survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$add_step(step)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step}}{The step to add.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-add_recipe"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-add_recipe}{}}}
\subsection{Method \code{add_recipe()}}{
Add a recipe to the survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$add_recipe(recipe, bake = lazy_default())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{recipe}}{The recipe to add.}

\item{\code{bake}}{Whether to bake the recipe immediately.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-add_workflow"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-add_workflow}{}}}
\subsection{Method \code{add_workflow()}}{
Add a workflow to the survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$add_workflow(workflow)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{workflow}}{The workflow to add.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-bake"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-bake}{}}}
\subsection{Method \code{bake()}}{
Bake all recipes in the survey.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$bake()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-head"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-head}{}}}
\subsection{Method \code{head()}}{
Display the first few rows of the survey data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$head()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-str"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-str}{}}}
\subsection{Method \code{str()}}{
Display the structure of the survey data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-set_design"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-set_design}{}}}
\subsection{Method \code{set_design()}}{
Set the survey design.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$set_design(design)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{design}}{The new survey design.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-update_design"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-update_design}{}}}
\subsection{Method \code{update_design()}}{
Update the survey design based on the current data and weights.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$update_design()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Survey-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Survey-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Survey$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
