% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_empty}
\alias{survey_empty}
\title{survey_empty}
\usage{
survey_empty(edition = NULL, type = NULL, weight = NULL, engine = NULL)
}
\arguments{
\item{edition}{The edition of the survey.}

\item{type}{The type of the survey.}

\item{weight}{The weight(s) of the survey.}

\item{engine}{The engine of the survey.}
}
\value{
An empty Survey object.
}
\description{
Create an empty Survey object.
}
\keyword{Surveymethods}
