% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_to_data.table}
\alias{survey_to_data.table}
\title{survey_to_data.table}
\usage{
survey_to_data.table(svy)
}
\arguments{
\item{svy}{Survey object}
}
\value{
data.table
}
\description{
Convert survey to data.table
}
