% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_to_tibble}
\alias{survey_to_tibble}
\title{survey_to_tibble}
\usage{
survey_to_tibble(svy)
}
\arguments{
\item{svy}{Survey object}
}
\value{
tibble
}
\description{
Convert survey to tibble
}
\keyword{Surveymethods}
