% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes.R
\docType{class}
\name{Token-class}
\alias{Token-class}
\alias{recipe}
\title{OAuth token objects.}
\format{
An R6 class object.
}
\description{
These objects represent the complete set of data needed for OAuth access:
an app, an endpoint, cached credentials and parameters. They should be
created through their constructor functions \code{\link[=oauth1.0_token]{oauth1.0_token()}}
and \code{\link[=oauth2.0_token]{oauth2.0_token()}}.
}
\section{Methods}{

\itemize{
\item \code{add()}: caches token to disk
\item \code{get_name()}: returns list of url and config
\item \code{print()}: refresh access token (if possible)
}
}

\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-recipe-new}{\code{recipe$new()}}
\item \href{#method-recipe-add}{\code{recipe$add()}}
\item \href{#method-recipe-get_name}{\code{recipe$get_name()}}
\item \href{#method-recipe-print}{\code{recipe$print()}}
\item \href{#method-recipe-clone}{\code{recipe$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recipe-new"></a>}}
\if{latex}{\out{\hypertarget{method-recipe-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recipe$new(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recipe-add"></a>}}
\if{latex}{\out{\hypertarget{method-recipe-add}{}}}
\subsection{Method \code{add()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recipe$add(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recipe-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-recipe-get_name}{}}}
\subsection{Method \code{get_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recipe$get_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recipe-print"></a>}}
\if{latex}{\out{\hypertarget{method-recipe-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recipe$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recipe-clone"></a>}}
\if{latex}{\out{\hypertarget{method-recipe-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{recipe$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
