% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_survey.R
\name{validate_recipe}
\alias{validate_recipe}
\title{Check valid recipe}
\usage{
validate_recipe(svy_type, svy_edition, recipe_svy_edition, recipe_svy_type)
}
\arguments{
\item{svy_type}{Type of survey}

\item{svy_edition}{Edition of the survey}

\item{recipe_svy_edition}{Edition of the recipe}

\item{recipe_svy_type}{Type of the recipe}
}
\value{
Logical
}
\description{
Check valid recipe
}
\keyword{internal}
