% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{view_graph}
\alias{view_graph}
\title{View graph}
\usage{
view_graph(svy, init_step = "Load survey")
}
\arguments{
\item{svy}{Survey object}

\item{init_step}{Initial step}
}
\value{
Graph
}
\description{
View graph
}
