% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeWorkflow.R
\name{workflow_from_list}
\alias{workflow_from_list}
\title{Construct a RecipeWorkflow from a plain list}
\usage{
workflow_from_list(lst)
}
\arguments{
\item{lst}{A list (typically from JSON) with workflow fields}
}
\value{
A RecipeWorkflow object
}
\description{
Construct a RecipeWorkflow from a plain list
}
