% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_replicate}
\alias{add_replicate}
\title{add_replicate}
\usage{
add_replicate(
  weight,
  replicate_pattern,
  replicate_path = NULL,
  replicate_id = NULL,
  replicate_type
)
}
\arguments{
\item{weight}{Weight}

\item{replicate_pattern}{Replicate pattern}

\item{replicate_path}{Replicate file}

\item{replicate_id}{Replicate ID}

\item{replicate_type}{Replicate type}
}
\description{
add_replicate
}
\keyword{utils}
