% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_weight}
\alias{add_weight}
\title{Add Weight time pattern}
\usage{
add_weight(monthly = NULL, annual = NULL, quarter = NULL)
}
\arguments{
\item{monthly}{Weight monthly}

\item{annual}{Weight annual}

\item{quarter}{Weight quarter}
}
\description{
Add Weight time pattern
}
