% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_list_workflows}
\alias{api_list_workflows}
\title{List workflows from API}
\usage{
api_list_workflows(
  search = NULL,
  survey_type = NULL,
  recipe_id = NULL,
  user = NULL,
  limit = 50,
  offset = 0
)
}
\arguments{
\item{search}{Text search}

\item{survey_type}{Filter by survey type}

\item{recipe_id}{Filter workflows that reference this recipe}

\item{user}{Filter by author email}

\item{limit}{Maximum results}

\item{offset}{Skip first N results}
}
\value{
List of RecipeWorkflow objects
}
\description{
Fetch workflows with optional search and filters.
}
\examples{
\dontrun{
api_list_workflows(survey_type = "ech")
}
}
\seealso{
Other api-workflows: 
\code{\link{api_get_workflow}()},
\code{\link{api_publish_workflow}()}
}
\concept{api-workflows}
