% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_logout}
\alias{api_logout}
\title{Logout}
\usage{
api_logout()
}
\value{
Invisibly, NULL.
}
\description{
Clear the stored API token from memory and the environment.
}
\examples{
\dontrun{
api_logout()
}
}
\seealso{
Other api-auth: 
\code{\link{api_login}()},
\code{\link{api_me}()},
\code{\link{api_refresh_token}()},
\code{\link{api_register}()},
\code{\link{configure_api}()}
}
\concept{api-auth}
