% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_publish_recipe}
\alias{api_publish_recipe}
\title{Publish a recipe}
\usage{
api_publish_recipe(recipe)
}
\arguments{
\item{recipe}{A Recipe object}
}
\value{
Invisibly, the API response with the assigned ID.
}
\description{
Publish a Recipe object to the API. Requires authentication
(call \code{api_login()} first).
}
