% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_client.R
\name{api_publish_workflow}
\alias{api_publish_workflow}
\title{Publish a workflow}
\usage{
api_publish_workflow(workflow)
}
\arguments{
\item{workflow}{A RecipeWorkflow object}
}
\value{
Invisibly, the API response.
}
\description{
Publish a RecipeWorkflow object to the API.
Requires authentication.
}
\examples{
\dontrun{
api_publish_workflow(my_workflow)
}
}
\seealso{
Other api-workflows: 
\code{\link{api_get_workflow}()},
\code{\link{api_list_workflows}()}
}
\concept{api-workflows}
