% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{bake_recipes}
\alias{bake_recipes}
\title{bake_recipes}
\usage{
bake_recipes(svy)
}
\arguments{
\item{svy}{A Survey object with recipes to bake.}
}
\value{
A Survey object with baked recipes.
}
\description{
Bake all recipes in a Survey object.
}
\keyword{Surveymethods}
