% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{bake_recipes}
\alias{bake_recipes}
\title{Bake recipes}
\usage{
bake_recipes(svy)
}
\arguments{
\item{svy}{Survey object}
}
\value{
Survey object with all recipes applied
}
\description{
Bake recipes
}
\examples{
\dontrun{
svy <- load_survey("data.csv", svy_type = "ech", svy_edition = "2023",
  recipes = my_recipe)
processed <- bake_recipes(svy)
}
}
\keyword{Surveymethods}
