% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{cat_design_type}
\alias{cat_design_type}
\title{cat_design_type}
\usage{
cat_design_type(self)
}
\arguments{
\item{self}{Object of class Survey}
}
\description{
Cast design type from survey
}
\keyword{Surveymethods}
