% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeCategory.R
\name{default_categories}
\alias{default_categories}
\title{Default recipe categories}
\usage{
default_categories()
}
\value{
List of RecipeCategory objects
}
\description{
Returns a list of standard built-in categories for recipe classification.
}
\examples{
cats <- default_categories()
vapply(cats, function(c) c$name, character(1))
}
