% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_engine.R
\name{default_engine}
\alias{default_engine}
\title{Set default engine}
\usage{
default_engine(.engine = "data.table")
}
\arguments{
\item{.engine}{Character vector with the name of the
default engine. By default, "data.table" is used.}
}
\description{
Sets a default engine for loading surveys.
If an engine is already configured, it keeps it;
otherwise, it sets "data.table" as the default engine.
}
\keyword{internal}
