% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaluate_cv}
\alias{evaluate_cv}
\title{Evaluate estimation with Coefficient of Variation}
\usage{
evaluate_cv(cv)
}
\arguments{
\item{cv}{Numeric coefficient of variation value.}
}
\value{
Character string with the quality category (e.g. "Excellent", "Good").
}
\description{
Evaluate estimation with Coefficient of Variation
}
\examples{
evaluate_cv(3) # "Excellent"
evaluate_cv(12) # "Good"
evaluate_cv(30) # "Use with caution"
}
\keyword{utils}
