% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_time_pattern}
\alias{extract_time_pattern}
\title{Extract time pattern}
\usage{
extract_time_pattern(svy_edition)
}
\arguments{
\item{svy_edition}{Survey edition}
}
\value{
List
}
\description{
Extract time pattern
}
\keyword{utils}
