% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_tidy_api.R
\name{find_workflows_for_recipe}
\alias{find_workflows_for_recipe}
\title{Find workflows that use a specific recipe}
\usage{
find_workflows_for_recipe(recipe_id)
}
\arguments{
\item{recipe_id}{Character recipe ID to search for.}
}
\value{
List of RecipeWorkflow objects that reference this recipe.
}
\description{
Cross-reference query: find all workflows that reference a given recipe ID.
}
\examples{
\dontrun{
wfs <- find_workflows_for_recipe("recipe_001")
}

}
\seealso{
\code{\link{filter_workflows}}
}
