% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_data}
\alias{get_data}
\title{get_data}
\usage{
get_data(svy)
}
\arguments{
\item{svy}{Survey object}
}
\value{
Data
}
\description{
Get data from survey
}
\examples{
dt <- data.table::data.table(id = 1:5, age = c(25, 30, 45, 50, 60), w = rep(1, 5))
svy <- Survey$new(data = dt, edition = "2023", type = "ech",
  psu = NULL, engine = "data.table", weight = add_weight(annual = "w"))
head(get_data(svy))
}
\keyword{Surveymethods}
