% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_engine.R
\name{get_engine}
\alias{get_engine}
\title{get_engine}
\usage{
get_engine()
}
\value{
Character vector with the name of the configured engine.
}
\description{
This function retrieves the currently configured engine for loading surveys. It returns the engine configured in the system options or environment variables.
}
\examples{
get_engine()
}
\keyword{engine}
