% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelSurvey.R
\name{get_follow_up}
\alias{get_follow_up}
\title{get_follow_up}
\usage{
get_follow_up(
  RotativePanelSurvey,
  index = 1:length(RotativePanelSurvey$follow_up)
)
}
\arguments{
\item{RotativePanelSurvey}{A RotativePanelSurvey object}

\item{index}{An integer}
}
\value{
A list
}
\description{
get_follow_up
}
\keyword{RotativePanelSurvey}
\keyword{Surveymethods}
