% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelSurvey.R
\name{get_implantation}
\alias{get_implantation}
\title{get_implantation}
\usage{
get_implantation(RotativePanelSurvey)
}
\arguments{
\item{RotativePanelSurvey}{A RotativePanelSurvey object}
}
\value{
A list
}
\description{
get_implantation
}
\keyword{RotativePanelSurvey}
\keyword{Surveymethods}
