% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{get_recipe}
\alias{get_recipe}
\title{Get a recipe from the API}
\usage{
get_recipe(svy_type = NULL, svy_edition = NULL, topic = NULL)
}
\arguments{
\item{svy_type}{A character string with the survey type of the recipe}

\item{svy_edition}{A character string with the survey edition of the recipe}

\item{topic}{A character string with the topic of the recipe}
}
\value{
A Recipe object
}
\description{
Get a recipe from the API
}
\keyword{Recipes}
\keyword{Survey}
\keyword{methods}
