% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lazy_default}
\alias{lazy_default}
\title{Lazy processing}
\usage{
lazy_default()
}
\value{
Logical indicating the current lazy processing setting.
}
\description{
Lazy processing
}
\examples{
# Check current lazy processing default
lazy_default()
}
\seealso{
Other options: 
\code{\link{get_engine}()},
\code{\link{set_engine}()},
\code{\link{set_lazy_processing}()},
\code{\link{set_use_copy}()},
\code{\link{show_engines}()},
\code{\link{use_copy_default}()}
}
\concept{options}
\keyword{utils}
