% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_survey.R
\name{load_survey}
\alias{load_survey}
\title{Load survey}
\usage{
load_survey(
  path = NULL,
  svy_type = NULL,
  svy_edition = NULL,
  svy_weight = NULL,
  ...
)
}
\arguments{
\item{path}{Path to the survey file}

\item{svy_type}{Type of survey}

\item{svy_edition}{Edition of the survey}

\item{svy_weight}{Weight of the survey}

\item{...}{Additional arguments}
}
\value{
Survey object
}
\description{
Load survey
}
\examples{
set_engine("data.table")
svy_example <- load_survey(
  svy_type = "eaii",
  svy_edition = "2019-2021",
  svy_weight = add_weight(annual = "w_trans"),
  input = "https://raw.githubusercontent.com/metasurveyr/metasurvey_data/main/eaii/2019-2021.csv",
  dec = ","
)
svy_example
}
\keyword{preprocessing}
