% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_survey.R
\name{load_survey}
\alias{load_survey}
\title{Load survey}
\usage{
load_survey(
  path = NULL,
  svy_type = NULL,
  svy_edition = NULL,
  svy_weight = NULL,
  svy_psu = NULL,
  ...,
  bake = FALSE
)
}
\arguments{
\item{path}{Path to the survey file}

\item{svy_type}{Type of survey}

\item{svy_edition}{Edition of the survey}

\item{svy_weight}{Weight of the survey}

\item{svy_psu}{Primary sampling unit}

\item{...}{Additional arguments}

\item{bake}{Logical}
}
\value{
Survey object
}
\description{
Load survey
}
\examples{
set_engine("data.table")
svy_example <- load_survey(
  load_survey_example(
     svy_type = "eaii",
     svy_edition = "2019-2021"
  ),
  svy_type = "eaii",
  svy_edition = "eaii_2019-2021",
  svy_weight = add_weight(annual = "w_trans"),
  dec = ","
)
svy_example
}
\keyword{preprocessing}
