% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaSurvey-package.R
\docType{package}
\name{metaSurvey-package}
\alias{metaSurvey}
\alias{metaSurvey-package}
\title{metaSurvey: Survey processing with metaprogramming}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Mauro Loprete \email{mloprete@cognus.cl}

Authors:
\itemize{
  \item COGNUS [copyright holder, funder]
  \item Natalia da Silva
}

}
\keyword{internal}
