% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{publish_recipe}
\alias{publish_recipe}
\title{Publish a recipe to the API}
\usage{
publish_recipe(recipe)
}
\arguments{
\item{recipe}{A Recipe object}
}
\value{
A JSON object
}
\description{
Publish a recipe to the API
}
\keyword{Recipes}
\keyword{Survey}
\keyword{methods}
