% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{read_recipe}
\alias{read_recipe}
\title{Load a recipe from a file}
\usage{
read_recipe(file)
}
\arguments{
\item{file}{A character string with the file path}
}
\value{
A Recipe object
}
\description{
Load a recipe from a file
}
