% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeWorkflow.R
\name{read_workflow}
\alias{read_workflow}
\title{Read a RecipeWorkflow from a JSON file}
\usage{
read_workflow(file)
}
\arguments{
\item{file}{Character file path}
}
\value{
A RecipeWorkflow object
}
\description{
Read a RecipeWorkflow from a JSON file
}
\examples{
\dontrun{
wf <- read_workflow("my_workflow.json")
}
}
\seealso{
Other workflows: 
\code{\link{RecipeWorkflow-class}},
\code{\link{evaluate_cv}()},
\code{\link{print.RecipeWorkflow}()},
\code{\link{publish_workflow}()},
\code{\link{reproduce_workflow}()},
\code{\link{save_workflow}()},
\code{\link{workflow}()},
\code{\link{workflow_from_list}()}
}
\concept{workflows}
\keyword{workflow}
