% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{recipe}
\alias{recipe}
\title{Recipe}
\usage{
recipe(...)
}
\arguments{
\item{...}{A list with the following metadata: name, user, svy, description}
}
\value{
A Recipe object
}
\description{
Recipe
}
\keyword{Recipes}
\keyword{Survey}
\keyword{methods}
