% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{save_recipe}
\alias{save_recipe}
\title{Save a recipe to a file}
\usage{
save_recipe(recipe, file)
}
\arguments{
\item{recipe}{A Recipe object}

\item{file}{A character string with the file path}
}
\description{
Save a recipe to a file
}
\keyword{Recipes}
\keyword{Survey}
\keyword{methods}
