% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{save_recipe}
\alias{save_recipe}
\title{Save Recipe}
\usage{
save_recipe(recipe, file)
}
\arguments{
\item{recipe}{A Recipe object.}

\item{file}{A character string specifying the file path.}
}
\value{
NULL.
}
\description{
Saves a Recipe object to a file in JSON format.
}
\details{
This function encodes the Recipe object and writes it to a JSON file.
}
\examples{
\dontrun{
# Example of saving a Recipe object
save_recipe(recipe_obj, "recipe.json")
}
}
\keyword{utils}
