% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeWorkflow.R
\name{save_workflow}
\alias{save_workflow}
\title{Save a RecipeWorkflow to a JSON file}
\usage{
save_workflow(wf, file)
}
\arguments{
\item{wf}{A RecipeWorkflow object}

\item{file}{Character file path}
}
\value{
NULL (called for side-effect)
}
\description{
Save a RecipeWorkflow to a JSON file
}
\examples{
\dontrun{
save_workflow(wf, "my_workflow.json")
}
}
\keyword{Workflows}
