% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeAPI.R
\name{set_backend}
\alias{set_backend}
\title{Set recipe backend}
\usage{
set_backend(type, path = NULL)
}
\arguments{
\item{type}{Character. "local" or "mongo".}

\item{path}{Character. File path for local backend.}
}
\description{
Configure the active recipe backend via options.
}
\examples{
\dontrun{
set_backend("local", path = "my_recipes.json")
set_backend("mongo")
}
}
