% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{set_data}
\alias{set_data}
\title{set_data}
\usage{
set_data(svy, data, .copy = use_copy_default())
}
\arguments{
\item{svy}{A Survey object.}

\item{data}{The new data.}

\item{.copy}{Whether to return a copy of the Survey object (default: TRUE).}
}
\value{
The updated Survey object.
}
\description{
Set new data for a Survey object.
}
\keyword{Surveymethods}
