% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_engine.R
\name{set_engine}
\alias{set_engine}
\title{set_engine}
\usage{
set_engine(.engine = show_engines())
}
\arguments{
\item{.engine}{Character vector with the name of the engine to configure. By default, the engine returned by the \code{show_engines()} function is used.}
}
\value{
Invisibly, the configured engine name.
}
\description{
This function configures the engine to be used for loading surveys. It checks if the provided engine is supported, sets the default engine if none is specified, and generates a message indicating the configured engine. If the engine is not supported, it throws an error.
}
\examples{
\dontrun{
set_engine("data.table")
}
}
\keyword{engine}
