% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_engine.R
\name{set_engine}
\alias{set_engine}
\title{set_engine}
\usage{
set_engine(.engine = show_engines())
}
\arguments{
\item{.engine}{Character vector with the name of the engine}
}
\description{
Set engine for load survey
}
