% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_lazy_processing}
\alias{set_lazy_processing}
\title{Set lazy processing}
\usage{
set_lazy_processing(lazy)
}
\arguments{
\item{lazy}{Lazy processing}
}
\description{
Set lazy processing
}
\keyword{utils}
