% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_lazy_processing}
\alias{set_lazy_processing}
\title{Set lazy processing}
\usage{
set_lazy_processing(lazy)
}
\arguments{
\item{lazy}{Logical. If TRUE, steps are deferred until bake_steps() is called.}
}
\description{
Set lazy processing
}
\examples{
old <- lazy_default()
set_lazy_processing(FALSE)
lazy_default() # now FALSE
set_lazy_processing(old) # restore
}
\keyword{utils}
