% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe_tidy_api.R
\name{set_user_info}
\alias{set_user_info}
\title{Set user info on a recipe}
\usage{
set_user_info(recipe, user)
}
\arguments{
\item{recipe}{A Recipe object.}

\item{user}{A RecipeUser object.}
}
\value{
The modified Recipe object.
}
\description{
Pipe-friendly function to assign a \code{\link{RecipeUser}} to a Recipe.
}
\examples{
\dontrun{
user <- recipe_user("Juan Perez", email = "juan@example.com")
my_recipe |> set_user_info(user)
}

}
\seealso{
\code{\link{recipe_user}}
}
