% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_engine.R
\name{show_engines}
\alias{show_engines}
\title{List available survey data engines}
\usage{
show_engines()
}
\value{
Character vector with the names of the available engines.
}
\description{
Returns a character vector of available
engines that can be used for loading surveys.
}
\examples{
show_engines()
}
\seealso{
Other options: 
\code{\link{get_engine}()},
\code{\link{lazy_default}()},
\code{\link{set_engine}()},
\code{\link{set_lazy_processing}()},
\code{\link{set_use_copy}()},
\code{\link{use_copy_default}()}
}
\concept{options}
