% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_compute}
\alias{step_compute}
\title{Step compute to create new variables}
\usage{
step_compute(
  svy = NULL,
  ...,
  .by = NULL,
  use_copy = use_copy_default(),
  comment = "Compute step",
  .level = "auto"
)
}
\arguments{
\item{svy}{\code{Survey} object or \code{RotativePanelSurvey}}

\item{...}{\verb{<data-masking>} Name-value pairs. The name gives the name of the column in the output.

The value can be:
\itemize{
\item A vector of length 1, which will be recycled to the correct length.
\item A vector the same length as the current group (or the whole data frame if ungrouped).
\item NULL, to remove the column.
A data frame or tibble, to create multiple columns in the output.
}}

\item{.by}{Strig with group variable to compute}

\item{use_copy}{Logical, if its FALSE adds the new variables to reference object like \code{data.table}, if it is TRUE adds the new variables to a new object}

\item{comment}{String to add metadata information to the stap compute compatible with Markdown syntax}

\item{.level}{String use for \code{Rotative_Panel_survey} object indicating data to use, level options available;  'implantation' or 'follow-up'}
}
\value{
\code{Survey}  or \code{Rotative_Panel_Survey} object
}
\description{
Function to create new variables similar to Stata COMPUTE function for the same porpose
}
\examples{
\dontrun{
step_compute(
pd=POBPCOAC \%in\% 3:5,
.default = 0,
comment = "Población Desocupada",
.level = "follow_up"
)
}
}
\keyword{Steps}
