% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_compute}
\alias{step_compute}
\title{Step compute}
\usage{
step_compute(svy, ..., use_copy = use_copy_default())
}
\arguments{
\item{svy}{Survey object}

\item{...}{Expressions to compute}

\item{use_copy}{Use copy}
}
\value{
Survey object
}
\description{
Step compute
}
