% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_compute}
\alias{step_compute}
\title{Step compute}
\usage{
step_compute(
  svy = NULL,
  ...,
  .by = NULL,
  use_copy = use_copy_default(),
  comment = "Compute step",
  .level = "auto"
)
}
\arguments{
\item{svy}{Survey object}

\item{...}{Expressions to compute}

\item{.by}{By}

\item{use_copy}{Use copy}

\item{comment}{Comment}

\item{.level}{Level}
}
\value{
Survey object
}
\description{
Step compute
}
\keyword{Steps}
