% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_recode}
\alias{step_recode}
\title{Step recode}
\usage{
step_recode(svy, new_var, ..., .default = NA_character_, .name_step = NULL)
}
\arguments{
\item{svy}{Survey object}

\item{new_var}{New variable}

\item{...}{Expressions to recode}

\item{.default}{Default value}

\item{.name_step}{Name of the step}
}
\value{
Survey object
}
\description{
Step recode
}
