% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_recode}
\alias{step_recode}
\title{Step recode}
\usage{
step_recode(
  svy = survey_empty(),
  new_var,
  ...,
  .default = NA_character_,
  .name_step = NULL,
  ordered = FALSE,
  use_copy = use_copy_default(),
  comment = "Recode step",
  .to_factor = FALSE,
  .level = "auto"
)
}
\arguments{
\item{svy}{Survey object}

\item{new_var}{New variable}

\item{...}{Expressions to recode}

\item{.default}{Default value}

\item{.name_step}{Name of the step}

\item{ordered}{Ordered}

\item{use_copy}{Use copy}

\item{comment}{Comment}

\item{.to_factor}{To factor}

\item{.level}{Level}
}
\value{
Survey object
}
\description{
Step recode
}
\keyword{Steps}
