% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{step_rename}
\alias{step_rename}
\title{Rename variables in survey data (step)}
\usage{
step_rename(
  svy,
  ...,
  mapping = NULL,
  .copy = use_copy_default(),
  comment = "Rename variables",
  use_copy = deprecated(),
  lazy = lazy_default(),
  record = TRUE
)
}
\arguments{
\item{svy}{A Survey or RotativePanelSurvey object}

\item{...}{Pairs in the form new_name = old_name (unquoted or character)}

\item{mapping}{A named character vector of the form
\code{c(new_name = "old_name")}.}

\item{.copy}{Whether to operate on a copy (default: use_copy_default())}

\item{comment}{Optional description for the step}

\item{use_copy}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use \code{.copy} instead.}

\item{lazy}{Internal. Whether to delay execution.}

\item{record}{Internal. Whether to record the step.}
}
\value{
Survey object with the specified variables
renamed (or queued for renaming).
}
\description{
Creates a step that renames variables in the survey data when baked.
}
\examples{
dt <- data.table::data.table(
  id = 1:5, age = c(25, 30, 45, 50, 60),
  w = rep(1, 5)
)
svy <- Survey$new(
  data = dt, edition = "2023", type = "ech",
  psu = NULL, engine = "data.table", weight = add_weight(annual = "w")
)
svy2 <- step_rename(svy, edad = age)
svy2 <- bake_steps(svy2)
"edad" \%in\% names(get_data(svy2)) # TRUE
}
\seealso{
Other steps: 
\code{\link{bake_steps}()},
\code{\link{get_steps}()},
\code{\link{step_compute}()},
\code{\link{step_join}()},
\code{\link{step_recode}()},
\code{\link{step_remove}()},
\code{\link{view_graph}()}
}
\concept{steps}
