% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipes.R
\name{steps_to_recipe}
\alias{steps_to_recipe}
\title{Convert a list of steps to a recipe}
\usage{
steps_to_recipe(steps)
}
\arguments{
\item{steps}{A list with the steps of the recipe}
}
\value{
A Recipe object
}
\description{
Convert a list of steps to a recipe
}
\keyword{Recipes}
\keyword{Survey}
\keyword{methods}
