% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Survey.R
\name{survey_empty}
\alias{survey_empty}
\title{survey_empty}
\usage{
survey_empty(edition = NULL, type = NULL, weight = NULL, engine = NULL)
}
\arguments{
\item{edition}{Edition of survey}

\item{type}{Type of survey}

\item{weight}{Weight of survey}

\item{engine}{Engine of survey}
}
\value{
Survey object
}
\description{
Create an empty survey
}
