% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_to_data.table}
\alias{survey_to_data.table}
\title{survey_to_data.table}
\usage{
survey_to_data.table(svy)
}
\arguments{
\item{svy}{Survey object}
}
\value{
data.table
}
\description{
Convert survey to data.table
}
\examples{
dt <- data.table::data.table(id = 1:5, age = c(25, 30, 45, 50, 60), w = rep(1, 5))
svy <- Survey$new(
  data = dt, edition = "2023", type = "ech",
  psu = NULL, engine = "data.table", weight = add_weight(annual = "w")
)
result <- survey_to_data.table(svy)
data.table::is.data.table(result) # TRUE
}
\keyword{Surveymethods}
