% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_to_data_frame}
\alias{survey_to_data_frame}
\title{survey_to_data_frame}
\usage{
survey_to_data_frame(svy)
}
\arguments{
\item{svy}{A Survey object.}
}
\value{
A data.frame containing the survey data.
}
\description{
Convert a Survey object to a data.frame.
}
\keyword{Surveymethods}
