% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_to_data_frame}
\alias{survey_to_data_frame}
\title{survey_to_data_frame}
\usage{
survey_to_data_frame(svy)
}
\arguments{
\item{svy}{Survey object}
}
\value{
data.frame
}
\description{
Convert survey to data.frame
}
