% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_to_tibble}
\alias{survey_to_tibble}
\title{survey_to_tibble}
\usage{
survey_to_tibble(svy)
}
\arguments{
\item{svy}{Survey object}
}
\value{
tibble
}
\description{
Convert survey to tibble
}
\examples{
\dontrun{
dt <- data.table::data.table(id = 1:5, age = c(25, 30, 45, 50, 60), w = rep(1, 5))
svy <- Survey$new(data = dt, edition = "2023", type = "ech",
  psu = NULL, engine = "data.table", weight = add_weight(annual = "w"))
tbl <- survey_to_tibble(svy)
class(tbl)
}
}
\keyword{Surveymethods}
