% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_time_pattern}
\alias{validate_time_pattern}
\title{Validate time pattern}
\usage{
validate_time_pattern(svy_type = NULL, svy_edition = NULL)
}
\arguments{
\item{svy_type}{Survey type}

\item{svy_edition}{Survey edition}
}
\value{
Logical

List
}
\description{
Validate time pattern
}
\keyword{utils}
