% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{Workflow}
\usage{
workflow(survey, ..., estimation_type = "monthly")
}
\arguments{
\item{survey}{Survey object}

\item{...}{Calls}

\item{estimation_type}{Estimation type}
}
\value{
Data
}
\description{
Workflow
}
\keyword{Surveymethods}
