% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeWorkflow.R
\name{workflow_from_list}
\alias{workflow_from_list}
\title{Construct a RecipeWorkflow from a plain list}
\usage{
workflow_from_list(lst)
}
\arguments{
\item{lst}{A list (typically from JSON) with workflow fields}
}
\value{
A RecipeWorkflow object
}
\description{
Construct a RecipeWorkflow from a plain list
}
\examples{
lst <- list(
  name = "example", user = "test", survey_type = "ech",
  edition = "2023", estimation_type = "svymean"
)
wf <- workflow_from_list(lst)
}
\seealso{
Other workflows: 
\code{\link{RecipeWorkflow-class}},
\code{\link{evaluate_cv}()},
\code{\link{print.RecipeWorkflow}()},
\code{\link{publish_workflow}()},
\code{\link{read_workflow}()},
\code{\link{reproduce_workflow}()},
\code{\link{save_workflow}()},
\code{\link{workflow}()},
\code{\link{workflow_table}()}
}
\concept{workflows}
